requirements.txt
django-crispy-forms
djangorestframework
django-rest-auth
django-allauth
django_countries
gunicorn==19.6.0
dj-database-url==0.5.0
whitenoise==4.0
psycopg2==2.8.3
